Keep time location data within Discord bot
