+++ discordbot/bot.py
@@ -11,10 +11,10 @@ from interactions.client.errors import CommandCheckFailure, CommandOnCooldown, M
 from interactions.models import ActivityType, Guild, GuildText, Intents, Member, Role
 
 import discordbot.commands
-from discordbot import command, error_handling
+from discordbot import command, error_handling, geonames
 from discordbot.shared import guild_id
 from magic import fetcher, multiverse, oracle, whoosh_write
-from shared import configuration, geonames, perf, repo
+from shared import configuration, perf, repo
 from shared import redis_wrapper as redis
 from shared.settings import with_config_file
